+++ src/main/java/net/replaceitem/scarpet/graphics/script/functions/GraphicsFunctions.java
@@ -1,7 +1,10 @@
 package net.replaceitem.scarpet.graphics.script.functions;
 
+import carpet.script.Context;
 import carpet.script.annotation.ScarpetFunction;
+import carpet.script.argument.FileArgument;
 import net.replaceitem.scarpet.graphics.script.values.PixelAccessibleValue;
+import org.apache.commons.lang3.tuple.Pair;
 import org.jspecify.annotations.Nullable;
 import processing.awt.PGraphicsJava2D;
 import processing.core.PConstants;
@@ -10,23 +13,36 @@ import processing.core.PImage;
 import net.replaceitem.scarpet.graphics.ScarpetGraphics;
 
 import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
+import java.net.URI;
 import java.net.URL;
+import java.util.Optional;
 
-@SuppressWarnings("unused")
+@SuppressWarnings({"unused", "OptionalUsedAsFieldOrParameterType"})
 public class GraphicsFunctions {
-    @ScarpetFunction
+    @ScarpetFunction(maxParams = 2)
     @Nullable
-    public PImage load_image(String location) {
+    public PImage load_image(Context c, String location, Optional<Boolean> shared) {
         try {
             try {
-                URL url = new URL(location);
+                URL url = URI.create(location).toURL();
                 return new PImage(ImageIO.read(url));
-            } catch (MalformedURLException e) {
-                File file = new File(location);
-                return new PImage(ImageIO.read(file));
+            } catch (MalformedURLException | IllegalArgumentException e) {
+                Pair<String, String> resource = FileArgument.recognizeResource(location, false, FileArgument.Type.ANY);
+                FileArgument fileArgument = new FileArgument(resource.getLeft(), FileArgument.Type.ANY, resource.getRight(), false, shared.orElse(false), FileArgument.Reason.READ, c.host);
+                if(c.host.main == null) return null;
+                @Nullable BufferedImage[] img = new @Nullable BufferedImage[]{null};
+                fileArgument.findPathAndApply(c.host.main, path -> {
+                    try {
+                        img[0] = ImageIO.read(path.toAbsolutePath().toFile());
+                    } catch (IOException ex) {
+                        throw new RuntimeException(ex);
+                    }
+                });
+                return new PImage(img[0]);
             }
         } catch (IOException e) {
             ScarpetGraphics.LOGGER.error(e);
@@ -34,9 +50,14 @@ public class GraphicsFunctions {
         }
     }
 
-    @ScarpetFunction
-    public void save_image(PixelAccessibleValue<?> image, String location) {
-        image.getGraphics().save(location);
+    @ScarpetFunction(maxParams = 3)
+    public void save_image(Context c, PixelAccessibleValue<?> image, String location, Optional<Boolean> shared) {
+        Pair<String, String> resource = FileArgument.recognizeResource(location, false, FileArgument.Type.ANY);
+        FileArgument fileArgument = new FileArgument(resource.getLeft(), FileArgument.Type.ANY, resource.getRight(), false, shared.orElse(false), FileArgument.Reason.READ, c.host);
+        if(c.host.main == null) return;
+        fileArgument.findPathAndApply(c.host.main, path -> {
+            image.getGraphics().save(path.toAbsolutePath().toString());
+        });
     }
 
     @ScarpetFunction