+++ changelog.md
@@ -1,1 +1,2 @@
 * Updated to Minecraft 26.1
+* Made `load_image` and `save_image` use the scarpet file system and no longer allow absolute paths.